Don't wrap a single listener id within a Set in cbMap (perf/mem optimization).
